значительное переделывание ServiceConfig

+++ Deployer/Core/ServiceConfig.cs
@@ -1,9 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.IO;
 using System.Linq;
+using System.Runtime.Remoting;
 using System.ServiceProcess;
 using System.Text;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
 namespace Deployer.Core
@@ -12,20 +15,57 @@ namespace Deployer.Core
     {
         private Service baseService;
 
-        private Dictionary<string,string> Versions { get; set; }// все версии пара- <версия, расположение>
-        private string currentVersion;// рабочая версия сервиса
+        public Dictionary<string, string> Versions { get; } = new Dictionary<string, string>();// все версии пара- <версия, расположение>
+        private string currentVersion;
+
+        public string CurrentVersion
+        {
+            get { return currentVersion;}
+            set
+            {
+                if (value != null)
+                    if (Versions.ContainsKey(value))
+                        currentVersion = value;
+                    else AddVersion(value, true);
+                else currentVersion = null;
+            }
+        } // рабочая версия сервиса
+
         public string ServiceName { get; set; }// название сервиса
         public string DeployPath { get; set; } // путь развертывания
         public string SourcePath { get; set; } // путь откуда будут браться обновления
+        public string ExeFile { get; set; }
         public bool IsMultiversion { get; set; }// много или 1 бекап
 
-        public ServiceConfig(Service service, string sourcePath)
+        public bool IsBuilded()
+        {
+            return currentVersion != null;
+        }
+
+        public ServiceConfig(string exePath, string sourcePath)
+        {
+            Service service=new Service(exePath);
+            baseService = service;
+            SourcePath = sourcePath;
+            Init();
+            if (!IsMultiversion) AddVersion("");
+            else AddVersion(service.Version);
+            CurrentVersion = service.Version;
+        }
+        public ServiceConfig(Service service, string sourcePath)// создание конфигурации сервиса service с уазанием источника обновлений sourcePath
         {
             baseService = service;
             SourcePath = sourcePath;
             Init();
-            currentVersion = baseService.Version;
+
+            if (!IsMultiversion) CurrentVersion = "";
+            else CurrentVersion = service.Version;
         }
+
+        //protected void RegisterService()
+        //{
+        //    if(IsMultiversion)
+        //}
         public void AddVersion(string version,bool setAsCurrent=true)
         {
             AddVersion(version,CreateVersionFolderName(version),setAsCurrent);
@@ -35,39 +75,54 @@ namespace Deployer.Core
         {
             try
             {
-                Versions.Add(version,path);
-                if (setAsCurrent) currentVersion = version;
+                Versions.Add(version, path);
+                if (setAsCurrent) CurrentVersion = version;
+                // можно добавить обновление
+            }
+            catch (ArgumentNullException e)
+            {
+                if(version==null)
+                    throw new ArgumentNullException("Версия не указана", e);
             }
             catch (ArgumentException e)
             {
-
-                throw new ArgumentException("Такая версия уже существует", e);
+                if (version != null)
+                    throw new ArgumentException($"Версия {version} уже существует", e);
+                else throw e;
             }
         }
+
         public string CreateVersionFolderName(string version)
         {
-            return DeployPath + "_" + version;
+            string path =version!=""? DeployPath + "_" + version:DeployPath;
+            return path;
         }
 
         protected void CopyFiles(string source, string dest, List<string> ignoreList = null)
         {
-            List<string> files = Directory.GetFiles(source).ToList();
+            source = source.Replace(@"\", @"\\")+"\\\\";
+            DirectoryInfo directory=new DirectoryInfo(source);
+            List<FileInfo> files = directory.GetFiles("*",SearchOption.AllDirectories).ToList();
             if (ignoreList != null)
             {
                 foreach (var ignoreFile in ignoreList)
                 {
-                    files.Remove(ignoreFile);
+                    //files.Remove(ignoreFile);
                 }
             }
-
-            foreach (string file in files)
+            foreach (var file in files)
             {
-                File.Copy(file, dest);
+                string name= new Regex(source).Replace(file.FullName,"");// имя файла в каталоге (включая подпапку) 
+                string newName = Path.Combine(dest, name);
+                string newPath= Path.GetDirectoryName(newName);
+                /*if (!Directory.Exists(newPath)) */Directory.CreateDirectory(newPath);
+                File.Copy(file.FullName, newName,true);
             }
         }
-        public void BackupService(string newVersion,List<string> ignoreFiles=null)
+        public string BackupService(string newVersion,List<string> ignoreFiles=null)
         {
             string backupFolder;
+            //if(!IsBuilded())throw new ArgumentException($"не найден собранный экземпляр. Выполните сборку перед созданием резервной копии");
             if (IsMultiversion)
             {
                 //AddVersion(newVersion);
@@ -77,25 +132,47 @@ namespace Deployer.Core
             {
                 backupFolder = DeployPath + "_backup";
             }
-            BackupService(newVersion,backupFolder,ignoreFiles);
+            return BackupService(newVersion,backupFolder,ignoreFiles);
         }
 
-        public void DeleteVersion(string version)
+        //public string BackupService(string oldVersion, string newVersion, string backupDirectory, List<string> ignoreFiles = null)
+        //{
+        //    string fromPath = Versions[oldVersion];
+        //    if (Versions.ContainsKey(newVersion) && !IsMultiversion) return ($"версия ({Versions}) уже существует");
+        //    if (!Directory.Exists(backupDirectory)) Directory.CreateDirectory(backupDirectory);
+        //    if (IsMultiversion) AddVersion(newVersion, false);
+        //    else if (!Versions.ContainsKey(newVersion)) AddVersion("backup");
+        //    CopyFiles(DeployPath, backupDirectory, ignoreFiles);
+        //    return "done";
+        //}
+
+        public void DeleteVersion(string delVersion)
         {
-            string path = Versions[version];
+            if(delVersion==currentVersion)throw new InvalidOperationException($"Удаление действующей версии недопустимо");
+            string path = Versions[delVersion];
             Directory.Delete(path,true);
-            Versions.Remove(version);
+            Versions.Remove(delVersion);
         }
-        public void BackupService(string newVersion, string backupDirectory, List<string> ignoreFiles = null)
+
+        public string BackupService(string newVersion,string sourceDirectory, string backupDirectory, List<string> ignoreFiles = null)
         {
-            if (Versions.ContainsKey(newVersion) && !IsMultiversion) throw new ArgumentException($"версия ({Versions}) уже существует");
-            if (!Directory.Exists(backupDirectory)) Directory.CreateDirectory(backupDirectory);
-            if(IsMultiversion) AddVersion(newVersion,false);
-            CopyFiles(DeployPath, backupDirectory, ignoreFiles);
+            if (Versions.ContainsKey(newVersion) && IsMultiversion) return ($"версия ({Versions}) уже существует");
+            Directory.CreateDirectory(backupDirectory);
+            if (IsMultiversion) AddVersion(newVersion, false);
+            else if (!Versions.ContainsKey(newVersion)) AddVersion("backup",false);
+            if(backupDirectory!=sourceDirectory)
+                CopyFiles(sourceDirectory, backupDirectory, ignoreFiles);
+            return "done";
+        }
+        public string BackupService(string newVersion, string backupDirectory, List<string> ignoreFiles = null)
+        {
+            return BackupService(newVersion, DeployPath, backupDirectory, ignoreFiles);
+            // генерим событие по добавлению версиии
+
         }
         protected void SetCurVersion(string version)
         {
-            if(Versions.ContainsKey(version))currentVersion = version;
+            if(Versions.ContainsKey(version))CurrentVersion = version;
             else AddVersion(version,true);
         }
 
@@ -111,24 +188,33 @@ namespace Deployer.Core
         }
         public void StopService()
         {
-            System.ServiceProcess.ServiceController ser = new System.ServiceProcess.ServiceController(ServiceName);
+            ServiceController ser = new System.ServiceProcess.ServiceController(ServiceName);
             ser.Stop();
         }
         public void StartService()
         {
-            System.ServiceProcess.ServiceController ser = new System.ServiceProcess.ServiceController(ServiceName);
+            ServiceController ser = new System.ServiceProcess.ServiceController(ServiceName);
             ser.Start();
         }
-
-
         
-        public void Build()
+        public void Build()// первоначальная сборка
         {
+            Directory.CreateDirectory(DeployPath);
             CopyFiles(SourcePath,DeployPath);
+            CurrentVersion = baseService.Version;
+            if (Versions.Count==0) AddVersion(CurrentVersion,true);
         }
-        public void Build(string sourcePath)
+        public void Build(string version) // сборка существующей версии
         {
-            CopyFiles(sourcePath, DeployPath);
+            Directory.CreateDirectory(DeployPath);
+            //string sourcePath = Versions[version];
+            if(SourcePath!=DeployPath)
+                CopyFiles(SourcePath, DeployPath);
+            if (!IsMultiversion)
+            {
+                CurrentVersion = "";
+            }
+            else CurrentVersion = version;
         }
 
         public string GetVersionBackupFolder(string version)
@@ -140,7 +226,8 @@ namespace Deployer.Core
             if(baseService==null) baseService=new Service();
             ServiceName = baseService.ServiceName;
             DeployPath = baseService.ServicePath;
-            Versions=new Dictionary<string, string>();
+            ExeFile = baseService.ExeName;
+            //Versions=new Dictionary<string, string>();
         }
         public List<Service> GetAllVersionServices()
         {
@@ -148,10 +235,23 @@ namespace Deployer.Core
             foreach (var version in Versions.Keys)
             {
                 var newService = (Service) baseService.Clone();
+                newService.ServicePath = Versions[version];
+                newService.ServiceName = ServiceName;
+                newService.ExeName = ServiceName;
                 newService.Version = version;
                 services.Add(newService);
             }
             return services;
         }
+
+        //public string GetCurrentVersion()
+        //{
+        //    return CurrentVersion;
+        //}
+
+        //public void SetCurrentVersion(string version)
+        //{
+        //    CurrentVersion = version;
+        //}
     }
 }
